ne2000: fix mcast lookup to truncate CRC to 32-bit even on x86_64. Prior to this fix, some CRC checksums would cause a segfault when compiled as 64-bit x86_64.

+++ src/hardware/ne2000.cpp
@@ -1237,7 +1237,7 @@ bx_ne2k_c::mcast_index(const void *dst)
       }
   }
 
-  return (crc >> 26u);
+  return (uint32_t)(crc >> 26u); /* WARNING: Caller directly uses our 6-bit return as index. If not truncated, will cause a segfault */
 #undef POLYNOMIAL
 }
